Refactor mocks
