+++ src/mock/consts.js
@@ -53,72 +53,6 @@ const EventType = Object.freeze({
   },
 });
 
-const cityEventTypes = Object.values(EventType)
-  .filter((type) => type.group === `transfer`)
-  .map((type) => type.code);
-
-/**
- * @readonly
- * @const {Array} Predefined list of available locations.
- */
-const LOCATIONS = [
-  {
-    name: `New York`,
-    type: `city`,
-    eventTypes: new Set(cityEventTypes),
-  },
-  {
-    name: `London`,
-    type: `city`,
-    eventTypes: new Set(cityEventTypes),
-  },
-  {
-    name: `Berlin`,
-    type: `city`,
-    eventTypes: new Set(cityEventTypes),
-  },
-  {
-    name: `Airport 1`,
-    type: `airport`,
-    eventTypes: new Set([`train`, `taxi`, `drive`]),
-  },
-  {
-    name: `Airport 2`,
-    type: `airport`,
-    eventTypes: new Set([`train`, `taxi`, `drive`]),
-  },
-  {
-    name: `Museum 1`,
-    type: `museum`,
-    eventTypes: new Set([`train`, `taxi`, `drive`, `sightseeing`]),
-  },
-  {
-    name: `Museum 2`,
-    type: `museum`,
-    eventTypes: new Set([`train`, `taxi`, `drive`, `sightseeing`]),
-  },
-  {
-    name: `Restaurant 1`,
-    type: `restaurant`,
-    eventTypes: new Set([`train`, `taxi`, `drive`, `restaurant`]),
-  },
-  {
-    name: `Restaurant 2`,
-    type: `restaurant`,
-    eventTypes: new Set([`train`, `taxi`, `drive`, `restaurant`]),
-  },
-  {
-    name: `Hotel 1`,
-    type: `hotel`,
-    eventTypes: new Set([`train`, `taxi`, `drive`, `check-in`]),
-  },
-  {
-    name: `Hotel 2`,
-    type: `hotel`,
-    eventTypes: new Set([`train`, `taxi`, `drive`, `check-in`]),
-  },
-];
-
 /**
  * Enum for time in millisecond.
  *
@@ -150,7 +84,6 @@ export {
   EventType,
   DurationType,
   KeyName,
-  LOCATIONS,
   activities,
   transfers,
 };
